Реализована обоработка данных о внутренних тегах (Tag) и тегах для сайта (Label)

+++ app/src/main/java/com/mandarinkafe/mandarin/menu/data/dto/ItemSize.kt
@@ -1,7 +1,7 @@
 package com.mandarinkafe.mandarin.menu.data.dto
 
 data class ItemSize(
-    val portionWeightGrams: Double,
+    val portionWeightGrams: Float,
     val prices: List<Price>,
     val buttonImageUrl: String,
 )
